libs/pyproject.toml
[project]
name = "libs"
version = "0.1.0"
description = "Shared libraries for connectome analysis"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.0.0",
    "pandas>=2.3.0",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.0",
    "scikit-learn>=1.6.0",
    "scipy>=1.13.0",
    "neuprint-python>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.build.targets.sdist]
include = [
    "*.py",
    "/tests",
]

[tool.hatch.metadata]
allow-direct-references = true

